Refactor ArgumentUtils into BundleUtils
