+++ AnkiDroid/src/main/java/com/ichi2/utils/BundleUtils.kt
@@ -49,4 +49,8 @@ object BundleUtils {
         }
         return getLong(key)
     }
+
+    inline fun <reified T> Bundle.getSerializableWithCast(key: String): T {
+        return getSerializable(key) as T
+    }
 }
